requirements.txt
streamlit>=1.18.1
python-dotenv>=0.21.0
openai>=0.27.0
requests>=2.28.1
google-api-python-client>=2.70.0
google-auth>=2.17.0
